FIX: 이미지 단건 삭제 시 실제 파일도 제거
기존에는 deleteImage가 DB 레코드만 삭제해 스토리지에 orphan 파일이
누적되던 문제 해결. 이미지 메타데이터를 조회해 경로를 얻은 뒤
파일을 먼저 삭제하고 DB 레코드를 정리하도록 변경. findByImageCode
조회 메서드와 SQL 추가.

+++ src/main/java/sisosolsol/greenfire/image/service/ImageService.java
@@ -92,7 +92,11 @@ public class ImageService {
     }
 
     public void deleteImage(Integer imageCode) {
-        // TODO: 이미지 경로를 조회한 후 파일 삭제
+        ImageDTO image = imageMapper.findByImageCode(imageCode);
+        if (image == null) {
+            return;
+        }
+        fileUploadUtil.deleteFile(image.getPath());
         imageMapper.deleteByImageCode(imageCode);
     }
